requirements.txt
eventlet>=0.33.0
Flask>=2.0.0
pymysql>=1.0.0
python-dotenv>=0.19.0
Flask-CORS>=3.0.10
Flask-SocketIO>=5.1.1
Werkzeug>=2.0.0
numpy>=1.19.0
xgboost>=1.5.0
joblib>=1.1.0
cryptography>=41.0.0